requirements.txt
click==8.0.3
Jinja2==3.1.2
paramiko==2.11.0
PyYAML==6.0
